Makefile and Source Directory Refactor
This commit refactors the makefile by using wildcards to get the source files, adding a SRC_DIR, BIN_DIR, and TARGET variable, and also reflects the folder structure of the `src` directory into the `bin` directory. This commit also renames the architecture `i686` to `x86_64`, along with the some other simple fixes along the way.