Loading the GDT
This commit implements the Global Descriptor Table very simply with a kernel code, kernel data, user code, and user data segment. The TSS is set to null because the scheduling and switching of tasks will be easier to implement in software rather than using hardware. A Makefile option has also been added to run the OS with bochs for more debugging capabilities.